sling

+++ source/Slingshot.hx
@@ -29,13 +29,13 @@ class Slingshot extends FlxSprite {
         powerBar.kill();
 
         // cooldown initialization
-		cooldown = new FlxSprite(-100, -100);
-        cdColor = new FlxColor();
-        cdColor.setRGBFloat(0.5, 0.5, 0.5, 0.75);
-		cooldown.exists = false;
-        cooldown.scrollFactor.set();
-        needCD = false;
-        cdLength = 45;
+		//cooldown = new FlxSprite(-100, -100);
+        //cdColor = new FlxColor();
+        //cdColor.setRGBFloat(0.5, 0.5, 0.5, 0.75);
+		//cooldown.exists = false;
+        //cooldown.scrollFactor.set();
+        //needCD = false;
+        //cdLength = 45;
 
         bulletLife = new Map<FlxSprite, Float>();
         // 100 bullets
@@ -59,17 +59,16 @@ class Slingshot extends FlxSprite {
 
     override public function update(elapsed:Float):Void 
     {
-        /*
+        
         if (FlxG.keys.justPressed.SPACE && !powerBar.alive && backpack.hasSlingshot) {
             powerBar.revive();
         }
-        else if (FlxG.keys.justReleased.SPACE && powerBar.alive && backpack.hasSlingshot) {
-        */
-        if (FlxG.keys.justPressed.SPACE && backpack.hasSlingshot && (!needCD)) {
+        else if (FlxG.keys.justReleased.SPACE && powerBar.alive && backpack.hasSlingshot) {       
+        //if (FlxG.keys.justPressed.SPACE && backpack.hasSlingshot && (!needCD)) {
             Main.LOGGER.logLevelAction(LoggingInfo.USE_SLINGSHOT, {coor: player.x + ", " + player.y});
             FlxG.sound.playMusic("assets/slingshotShoot.ogg", 1, false);
-            // lifeSpan = powerBar.generateResult();
-            lifeSpan = 0.85;
+            lifeSpan = powerBar.generateResult();
+            //lifeSpan = 0.85;
             // Space bar was pressed! FIRE A BULLET
 			var bullet:FlxSprite = playerBullets.recycle();
 			bullet.reset(player.x + 16 - bullet.width/2, player.y);
@@ -85,24 +84,24 @@ class Slingshot extends FlxSprite {
 			else if (player.facing == FlxObject.DOWN) {
 					bullet.velocity.y = 400;
 			}
-            /*
+            
             if (lifeSpan <= 0) {
                 lifeSpan = 1/10;
             }
-            */
+            
             bulletLife[bullet] = lifeSpan;
-            needCD = true;
-            cdLength = 45;
-            cdIndex = backpack.currentItemIdx;
-            cooldown.makeGraphic(45, 45, cdColor);
-            var cdX = FlxG.camera.x + 165 + 45 * cdIndex;
-            var cdY = FlxG.camera.y + 435;
-            cooldown.reset(cdX, cdY);
+            //needCD = true;
+            //cdLength = 45;
+            //cdIndex = backpack.currentItemIdx;
+            //cooldown.makeGraphic(45, 45, cdColor);
+            //var cdX = FlxG.camera.x + 165 + 45 * cdIndex;
+            //var cdY = FlxG.camera.y + 435;
+            //cooldown.reset(cdX, cdY);
         } 
-        /* else if (!backpack.hasLasso && !backpack.hasSlingshot) {
+         else if (!backpack.hasLasso && !backpack.hasSlingshot) {
             powerBar.kill();
         }
-        */
+        /*
         else if (needCD) {
             cdLength--;
             cooldown.makeGraphic(45, cdLength, cdColor);
@@ -112,8 +111,8 @@ class Slingshot extends FlxSprite {
             if (cdLength == 1) {
                 needCD = false;
                 cooldown.kill();
-            }
-        }
+            } 
+        } */
 
         for (bullet in bulletLife.keys()) {
             var time = bulletLife[bullet];
